Armature Generation alpha
I'm not currently generating the armature properly, it's coming out as a bunch of separate objects, rather than bones of a single armature, which leads to a number of issues. However, as a proof of concept I'm pleased. I'm still working out how to build the armature properly, and I'll probably tackle the hands once I've got the body worked out. Then work out some script for baking to animation, or figure out how to retarget it to a proper armature, ideally both rigify and mr mannequin.