Rig scale, Anim Retiming, Marker sync, UI update
Scale the mocap rig to fall in line with Rigify and Mr Mannequin's skeletons, scale-wise, which should make future retargeting a bit easier.
Added a function to adjust the timing of the animation playback. My laptop can only capture the full rig at around 7fps, so, when it played back at 24fps it would be too fast. This function takes the fps measurement generated during capture, and divides the current frame rate by it in order to generate the retiming factor.
The markers for the face, hands and body previously moved independently of one another, they have now been synced via object constraints that link their roots to the main pose root in the x and z axes, and for the y axis, the face draws from the nose tracker, and the left and right hands draw from the left and right wrists, respectively.
UI has been updated to be more compact, removing some of the labels to preserve vertical space. The Retime Animation operator has been added to the GUI, and there have been small naming tweaks, and icon changes.