New Skeleton
Armature Generator now gives a proper armature with the appropriate bone hierarchy.
The bones are bound by bone constraints to the trackers. This gives good fidelity with the tracking, but it results in bones that move and stretch.
Initial retargeting tests show that more will need to be done in terms of stabilizing the data to make it usable.
My initial thoughts are to either try a camera tracking solver solution to properly place the trackers within 3D space, another thought was to create a second armature with more constraints to make up for some of the deficiencies in the tracked data.