Hand Armatures, tracker cleanup
The armature button has been renamed, and it now generates armatures for the left and right hand, if they are in the scene.
After generating the armature it now hides the empties, which serves to hide all of the "parent-child" dotted lines that clogged up the viewport, and I also chose to hide some of the superfluous trackers: the body tracking model comes with index, pinky and thumb for each hand, as well as a few face trackers that don't seem to work as well as the dedicated face and hands.

All of this also exposes the biggest weakness of Mediapipe, its Holistic model only stitches together three separate models that don't work particularly well together. The finger tracking for the body model will not exactly match the finger tracking from the hands, and ditto for the face, leading to jitter and drift, not only between the performance and the capture, but between the different capture models. I think an intermediate rig may be necessary, unless there is some way to constrain the current skeleton and preserve reasonable motion data fidelity.